backend/deprecatedpycode/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
pyjwt==2.8.0
pandas==2.1.1
bleak==0.21.1
python-dotenv==1.0.0

